+++ imod/mf6/interfaces/ipackage.py
@@ -26,3 +26,7 @@ class IPackage(IPackageBase, metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def auxiliary_data_fields(self) -> dict[str, str]:
         raise NotImplementedError
+
+    @abstractmethod
+    def is_regridding_supported(self) -> bool:
+        raise NotImplementedError