+++ imod/mf6/interfaces/isimulation.py
@@ -1,4 +1,5 @@
 from imod.mf6.interfaces.idict import IDict
+from imod.mf6.interfaces.imodel import IModel
 
 
 class ISimulation(IDict):
@@ -10,4 +11,7 @@ class ISimulation(IDict):
         raise NotImplementedError
 
     def has_one_flow_model(self) -> bool:
+        raise NotImplementedError
+
+    def get_models(self) -> dict[str, IModel]:
         raise NotImplementedError