+++ imod/mf6/utilities/regrid.py
@@ -309,6 +309,11 @@ def _regrid_like(
     a model with similar packages to the input model, and with all the data-arrays regridded to another discretization,
     similar to the one used in input argument "target_grid"
     """
+    supported, error_with_object_name = model.is_regridding_supported()
+    if not supported:
+        raise ValueError(
+            f"regridding this model cannot be done due to the presence of package {error_with_object_name}"
+        )
     new_model = model.__class__()
 
     for pkg_name, pkg in model.items():
@@ -366,6 +371,15 @@ def _regrid_like(
         raise ValueError(
             "Unable to regrid simulation. Regridding can only be done on simulations that have a single flow model."
         )
+
+    models = simulation.get_models()
+    for model_name, model in models.items():
+        supported, error_with_object_name = model.is_regridding_supported()
+        if not supported:
+            raise ValueError(
+                f"Unable to regrid simulation, due to the presence of package '{error_with_object_name}' in model {model_name} "
+            )
+
     result = simulation.__class__(regridded_simulation_name)
     for key, item in simulation.items():
         if isinstance(item, IModel):
@@ -404,6 +418,11 @@ def _regrid_like(
     """
     target_grid_2d = target_grid.isel(layer=0, drop=True, missing_dims="ignore")
     return clip_by_grid(package, target_grid_2d)
+
+
+@typedispatch  # type: ignore[no-redef]
+def _regrid_like(package: object, target_grid: GridDataArray, *_) -> None:
+    raise TypeError("this object cannot be regridded")
 
 
 def _get_regridding_domain(