+++ docs/api/changelog.rst
@@ -49,6 +49,7 @@ Added
 - added logging to various initialization methods, write methods and dump methods. `See the
   documentation <https://deltares.github.io/imod-python/api/generated/logging/imod.logging.html>`_
   how to activate logging.
+- simulations and models that include a lake package now raise an exception on clipping, partitioning or regridding. 
 
 Changed
 ~~~~~~~